feat(kernel): complete SPI unification, persistence gaps, and UI services
## Agent Kernel SPI Improvements

### Unified Provider Interface (new: provider.rs)
- Add AgentProvider super-trait with provider_manifest(), health(), capabilities()
- Add Cancellable trait for cancel-capable operations
- Add Streaming trait for streaming-capable providers
- Add PolicyGated trait for policy-aware operations
- Add Listable<T> trait for list/get/exists operations
- Add Lifecycle<C,U> trait for CRUD lifecycle management
- Add ProviderError enum with typed error variants
- Add ProviderRegistration and ProviderSource for runtime registration

### HostProvider Enhancements (host.rs)
- Add StorageResult, TimeResult, EnvironmentResult, ExecutorResult types
- Add storage(), time(), environment(), executor() trait methods with defaults
- Default implementations return CapabilityMissing for optional operations

### ContextProvider Enhancements (context_memory.rs)
- Add ContextRanking and ContextExplanation types
- Add rank(), trim(), explain() trait methods with default implementations
- Default rank() returns decreasing relevance scores
- Default trim() is pass-through (no-op)
- Default explain() returns generic explanation

### PlanningProvider Fix (planning.rs)
- Change create_plan() return type from Plan to KernelResult<Plan>
- Enables error propagation during plan creation

### PolicyProvider Enhancements (policy.rs)
- Add PolicyExplanation type
- Add explain() and record_decision() trait methods with defaults
- Default explain() extracts from safe_reason/reason_code
- Default record_decision() is no-op

### McpProvider Fix (mcp.rs)
- Change list_servers() return type from Vec to KernelResult<Vec>
- Enables error reporting for network-connected MCP servers

## Code Kernel Improvements

### Shared Policy Utils (new: policy_utils.rs)
- Extract first_policy_category() to shared module
- Remove duplicate implementations from workspace.rs, vcs.rs, patch.rs, terminal.rs

### VerificationPlan Enhancement (verification.rs)
- Add policy_categories field for consistency with other request types
- Add with_policy_categories() builder method

### KnowledgeSearchResult Fix (knowledge.rs)
- Change score type from u32 to f32 for fractional relevance scoring

## Agent Business Persistence Gaps

### PostgreSQL Schema (agent_business_postgres.sql)
- Add 'runtime_executed' to audit action CHECK constraint

### Domain Layer (domain.rs)
- Add 15 memory lifecycle audit action variants
- MemoryProfile Updated/Deleted/Restored
- MemoryBinding Updated/Deleted/Restored
- MemoryNamespace Updated/Deleted/Restored
- MemorySource Deleted/Restored
- MemoryRelation Deleted/Restored

### Ports Layer (ports.rs)
- Add 7 repository trait methods with defaults:
  list_memory_stores, update_memory_profile/binding/namespace,
  get_memory_source/relation/retrieval_index

### Application Layer (application.rs)
- Add 18 service methods for memory lifecycle CRUD
- All methods follow authorize->validate->execute->audit pattern

### Persistence Layer (persistence.rs)
- Add 8 SQL constants for memory operations

### Infrastructure Layer (infrastructure.rs)
- Implement 7 new repository methods in InMemoryAgentRepository
- Fix list_memory_sources and list_memory_relations sorting

## Kernel UI Services

### Real KernelUiClient (new: kernel-ui.real.ts)
- Add HttpKernelUiClient with baseUrl configuration
- Support loadSnapshot() and decidePermission() via HTTP
- Update index.ts to export both mock and real clients

### Shell App (App.tsx)
- Add environment-based client selection via VITE_KERNEL_API_URL
- Add vite-env.d.ts for Vite type declarations

## Test Results
- sdkwork-agent-kernel: 227 tests passing
- sdkwork-code-kernel: 33 tests passing
- sdkwork-agent-business: 122 tests passing
- sdkwork-kernel-ui: typecheck + architecture check passing

+++ sdkwork-code-kernel/src/knowledge.rs
@@ -108,11 +108,11 @@ impl KnowledgeQuery {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct KnowledgeSearchResult {
     pub document_id: String,
     pub kind: KnowledgeDocumentKind,
-    pub score: u32,
+    pub score: f32,
     pub title: String,
 }
 
@@ -120,7 +120,7 @@ impl KnowledgeSearchResult {
     pub fn new(
         document_id: impl Into<String>,
         kind: KnowledgeDocumentKind,
-        score: u32,
+        score: f32,
         title: impl Into<String>,
     ) -> Self {
         Self {
